♻️ refactor command definition for packetdiag
