+++ src/main/java/net/sourceforge/plantuml/packetdiag/command/CommandScaleDirection.java
@@ -33,41 +33,44 @@
  *
  * 
  */
-package net.sourceforge.plantuml.packetdiag;
+package net.sourceforge.plantuml.packetdiag.command;
+
+import java.util.Optional;
 
 import net.sourceforge.plantuml.command.CommandExecutionResult;
 import net.sourceforge.plantuml.command.ParserPass;
 import net.sourceforge.plantuml.command.SingleLineCommand2;
 import net.sourceforge.plantuml.klimt.color.NoSuchColorException;
+import net.sourceforge.plantuml.packetdiag.PacketDiagram;
 import net.sourceforge.plantuml.regex.IRegex;
 import net.sourceforge.plantuml.regex.RegexConcat;
 import net.sourceforge.plantuml.regex.RegexLeaf;
 import net.sourceforge.plantuml.regex.RegexResult;
 import net.sourceforge.plantuml.utils.LineLocation;
 
-public class CommandScaleInterval extends SingleLineCommand2<PacketDiagram> {
+public class CommandScaleDirection extends SingleLineCommand2<PacketDiagram> {
 
-	public CommandScaleInterval() {
+	public CommandScaleDirection() {
 		super(getRegexConcat());
 	}
 
 	static IRegex getRegexConcat() {
-		return RegexConcat.build(CommandScaleInterval.class.getName(), RegexLeaf.start(), //
-						new RegexLeaf("scale_interval"), //
+		return RegexConcat.build(CommandScaleDirection.class.getName(), RegexLeaf.start(), //
+						new RegexLeaf("scale_direction"), //
 						RegexLeaf.spaceZeroOrMore(), //
 						new RegexLeaf("="), //
 						RegexLeaf.spaceZeroOrMore(), //
-						new RegexLeaf(1, "INTERVAL", "(\\d+);?"), //
+						new RegexLeaf(1, "DIR", "(ltr|rtl);?"), //
 						RegexLeaf.spaceZeroOrMore(), RegexLeaf.end());
 	}
 
 	@Override
 	protected CommandExecutionResult executeArg(PacketDiagram system, LineLocation location, RegexResult arg, ParserPass currentPass) throws NoSuchColorException {
-		try {
-			system.updateScaleInterval(Integer.parseInt(arg.get("INTERVAL", 0)));
-		} catch (NumberFormatException e) {
-			return CommandExecutionResult.error("Scale interval invalid", e);
-		}
+		PacketDiagram.ScaleDirection dir = Optional.ofNullable(arg.get("DIR", 0))
+						.map(String::toUpperCase)
+						.map(PacketDiagram.ScaleDirection::valueOf)
+						.orElse(PacketDiagram.ScaleDirection.LTR);
+		system.setScaleDirection(dir);
 		return CommandExecutionResult.ok();
 	}
 }
